[BACKPORT 2024.2][PLAT-15488]: Move the logic to compute the pitr snapshot inteval to BE.
Summary:
Moved the logic to set PITR snapshot interval in BE and honour it if passed in API.
We would only set snapshot time if it is not passed in the payload and let it pass through our validation but if if it is not passed in the API payload we would calculate it using Math.min(yb.xcluster.transactional.pitr.default_snapshot_interval, pitr retention time - 1 second).

Also,  updated the default PITR snapshot from 1h to 1d.

Original diff/commit: 378b5e16f2/D38558

Test Plan:
Tested manually that snapshot is updated using the above formula during create/edit.
Will cover this in the upcoming edit PITR params local provider test

Reviewers: hzare, sanketh, cwang, jmak, sneelakantan, dshubin

Reviewed By: cwang

Subscribers: yugaware, dshubin

Differential Revision: https://phorge.dev.yugabyte.com/D38928