[PLAT-15523] Show tables tab for GET xCluster config without sync db option
Summary:
If at least one of the xCluster config participant universes is down, then prior to this diff we would not show the xCluster table list as table details information would be missing.
This diff updates the YBA UI such that we are now able to see the tables with the understanding that the table name will be replaced with table UUID, db field will be missing and table status will be hidden.
Table level lag graph can be shown as well, but likewise the table details will be missing so we only show table UUID.

We are making this change simply to give the user more information to work with when they encounter this situation from the YBA UI.

Test Plan:
Set up an xCluster config and stop all the nodes in one of the universes.
Verify that the xCluster table list is shown and is as described above.
{F293602}
{F293603}

Reviewers: rmadhavan, hzare, cwang, vbansal, kkannan

Reviewed By: kkannan

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D38523