[PLAT-15556] Pause universe ephemeral storage check should only apply to AWS universes
Summary:
The fix for https://yugabyte.atlassian.net/browse/PLAT-9803 identified AWS ephemeral
instances as anything with a letter 'd' in them. This check is being used to decide if pause
universe needs to be shown, so this diff fixes that check to apply to AWS univs only.

Test Plan: Create an azure universe with Standard Dsv3, verify universe is pausable.

Reviewers: jmak

Reviewed By: jmak

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D38596