[BACKPORT pg15-cherrypicks][#23686] YSQL: Build relcache foreign key list from YB catcache
Summary:
relcache.c:

  - Includes:
    - YB master 3e93354184529cbee6df27337b5a41d78e7eced0 includes  "utils/catcache.h"
    - YB pg15 55782d561e55ef972f2470a4ae887dd791bb4a97 adds includes for partcache.h and relcache.h
    - Keep both sets of includes
  - RelationGetFKeyList
    - YB master 3e93354184529cbee6df27337b5a41d78e7eced0 moves some code that uses heap_open/heap_close into ELSE blocks
    - YB pg15 initial merge 55782d561e55ef972f2470a4ae887dd791bb4a97 touches the same pieces of code and replaces heap_open/heap_close with table_open/table_close
    - Resolve by doing both: switching to table_open/table_close and moving into ELSE blocks
guc.h:

  - GUC declarations:
    - YB pg15 55782d561e55ef972f2470a4ae887dd791bb4a97 adds PGDLLIMPORT to GUC declarations.
    - YB master 3e93354184529cbee6df27337b5a41d78e7eced0 adds new GUC yb_enable_fkey_catcache
    - Resolve similar to YB pg15 initial merge 5627af5d6009b6d5b14bbbb552f6fd2a9ae43c3a by keeping new GUC with PGDLLIMPORT and moving yb_enable_fkey_catcache to where the other YB declarations are.
pg_catalog_perf-test.cc:
  - ForeignKeyRelcachePreloadTest:
    - pg15 makes changes to query planning, causing additional catalog lookup RPCs. Adjust the expected # of catalog RPCs to account for this.

Original commit: 3e93354184 / D37595
Today, when preloading is enabled (either `ysql_catalog_preload_additional_tables` or `ysql_catalog_preload_additional_table_list` is set), we preload the relcache, meaning that we load the relcache entry for every relation in one shot. When we do relcache preloading, we currently don't load foreign key lists—instead, we loading them on-demand. This means on a given connection, the first time each table with a foreign key is referenced, that connection has to fetch the foreign key list from the master leader.

With this revision, we make two changes:
(1) Modify `RelationGetFKeyList` so that it reads foreign keys from the Yugabyte-only `pg_constraint` catcache `YBCONSTRAINTRELIDTYPIDNAME`.  Each time `RelationGetFKeyList` is invoked, it will do a `SearchCatCacheList` on this cache with the partial key `conrelid`.
(2) Whenever we preload the `pg_constraint` cache, we also preload the list cache for the partial key `conrelid`. This has the effect that any calls to `RelationGetFKeyList` after preloading will be able to read from the cached list and will not have to go to master. (A lookup for a table created on a different connection after preloading will still need to go to master).

Both of these changes are controlled by a new GUC `yb_enable_fkey_catcache` (+tserver gflag wrapper) that is on by default.
Jira: DB-12593

Test Plan:
```
./yb_build.sh --cxx-test pg_catalog_perf-test --gtest_filter "*ForeignKeyRelcachePreloadTest*"
```

Reviewers: tfoucher, fizaa, telgersma

Reviewed By: fizaa

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D38511