[BACKPORT pg15-cherrypicks][#24123] YSQL: Make a new ANN Access Method that redirects to the base table
Summary:
Original commit: b9b57c6f791cb9f28d8bec9372e9a433a903980c / D38176

Modifications/Merge Conflicts:
- catalog/index.c
  - Upstream commit d12bdba77b0fce9df818bc84ad8b1d8e7a96614b moves the code to reindex all indexes within `reindex_relation`. This moved code in YB had a PK check to see whether or not the underlying DocDB index should be dropped. D38176's modification of that check resulted in a merge conflict.
  - Resolution: replace `iRel->rd_index->indisprimary` in `reindex_relation` with `!YBIsCoveredByMainTable(iRel)`

- indexcmds.c
  - D38176 makes sure to skip a call to `YBCPgBackfillIndex` if the mentioned index is "covered" by the main table via a condition check on `YBIsCoveredByMainTable`. 0fcf6e4ed48581de3cec14a77fd7662fdb4fd2a7/D28432 moves this line elsewhere within a check to `IsYugaByteEnabled()`. This change makes sure to move the covered check to the new location of `YBCPgBackfillIndex`.

- pg_yb_utils.c
 - `rd_amroutine` was changed to `rd_indam` by upstream commit 346ed70b0ad7efc574711a97812692dab4542712.

This change introduces infrastructure for creating indexes that are "covered" by the main table much like a primary key index. This is achieved by adding a boolean variable to the `IndexAmRoutine` struct by the name of `yb_amiscoveredbymaintable` to denote whether or not the index is inlined with the underlying table. YB code that has traditionally checked `indisprimary` to avoid doing operations to the main table is replaced by checks to a utility function `YBIsCoveredByMainTable(Relation)` that checks `indisprimary` along with `yb_amiscoveredbymaintable`.

This work is done in preparation for vector indexes that are colocated with an underlying table. This change introduces the index access method `yblocalann` as an example of such an index type. This index AM can only be created on colocated tables for now as there isn't infrastructure to read ordered results from a multi-tablet vector index yet.

**Upgrade/Rollback safety:**
This change moves where `PgVectorReadOptionsPB` is defined and adds a protobuf field to that `vector_column_id` to signify what column is being used as the vector key. This shouldn't affect any user of the previous format. This is also for a use case that is not GA or EA.
Jira: DB-13014

Test Plan:
Jenkins: rebase: pg15-cherrypicks

./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressThirdPartyExtensionsPgvector'

Reviewers: fizaa, telgersma

Reviewed By: telgersma

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D38556