[BACKPORT 2024.1][#24160] docdb: Remove table name equality check in restore path
Summary:
Original commit: d6210df158efe504b40e7d08807c872f8ff7a8e2 / D38512
D22774 introduced a check that the name of the table being restored and the name of the
table in the snapshot are the same. This is not true for PITR / clone if the table has been renamed
since the snapshot. This diff removes that check.

This might cause backups that were created with partial indexes and before diff D23486 to
restore with an incorrect schema if there is also a colocation id collision between the created
table and a table in the snapshot.
Jira: DB-13047

Test Plan:
Set `ysql_enable_packed_row_for_colocated_table = true` and run:
`./yb_build.sh release --cxx-test tools_yb-admin-snapshot-schedule-test --gtest_filter ColocationAndRestoreType/YbAdminSnapshotScheduleTestWithYsqlColocationRestoreParam.PgsqlRenameTable/4`

The test previously failed because we were skipping schema merging.

Reviewers: yguan, zdrudi

Reviewed By: zdrudi

Subscribers: ybase

Differential Revision: https://phorge.dev.yugabyte.com/D38635