[PLAT-11305] Adding prometheus log file
Summary: We were not passing the prometheus log file anywhere to actually write the logs but were showing it in the status output which was kind of confusing. This diff adds the StandardOutput and StandardError redirection fields directly to the service file. NOTE: This won't work on linux with systemd < 240 (centos 7/al2/etc) and it also will completely  move prometheus logging from journalctl so journalctl -u prometheus won't show the logs. Getting them to stay in both was a bit more challenging and couldn't get it to work.

Test Plan: install yba and make sure prometheus logs present

Reviewers: dshubin, sanketh

Reviewed By: dshubin

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D38644