[#24044] docdb: Add an RPC to remove tservers from the registry.
Summary:
Adds an RPC to the cluster service to remove tservers from the registry. The preconditions are:
  - the tserver is not heartbeating to the master leader
  - the tserver is not hosting any tablets
  - the tserver is blacklisted

Removing tablet servers from the registry has long been an ask from YBA. Furthermore because we're going to start persisting tablet server descriptors to the sys catalog, offering an API to remove them before we roll out persistence could be helpful if the feature goes sideways. A `yb-admin` command will be added in an upcoming diff.
Jira: DB-12935

Test Plan:
```
./yb_build.sh --with-tests --cxx-test-filter-re master_cluster_test --cxx-test master_cluster_test
```

Reviewers: asrivastava

Reviewed By: asrivastava

Subscribers: nsingh, sanketh, esheng, ybase, yql, ycdcxcluster, slingam

Differential Revision: https://phorge.dev.yugabyte.com/D38493