[#22519] YSQL: Enable batching of explicit row lock requests by default
Summary:
The main goal of this change is to enable batching of the explicit row locks by default. Desired behavior is achieved by changing the default value of the `yb_explicit_row_locking_batch_size` flag from `1` to `1024`.

The main changes in the diff relates to showing proper error message in case the lock can't be taken with `NOWAIT` policy. Native postgres message contains the table name and it looks this

```
ERROR:  could not obtain lock on row in relation "t"
```

In case of batching it is possible that YSQL will try to lock rows from multiple tables in one request. And it has to select exact table where locking failed.

Sample query with locking rows from multiple tables:

```
SELECT * FROM t1 INNER JOIN t2 ON t1.k = t2.k WHERE t1.k = 1 FOR SHARE NOWAIT
```

In the example above it is possible that locking of rows from `t1` conflicts with another transaction or rows from `t2` or both at same time.
So YSQL has to show the
```
ERROR:  could not obtain lock on row in relation "t1"
```
error message or
```
ERROR:  could not obtain lock on row in relation "t2"
```

Desired behavior is achieved by adding the `OpIndex` payload with index of operation caused conflict error to the `Status` object provide by the `CombineErrorsToStatus` function.

**Note:**
In case of conflict with another transaction for both tables `t1` and `t2` it is not possible to provide deterministic error message like postgres does. YSQL may generate error message for `t1` or `t2` with some probability.

The reason of difference in behavior of postgres and YSQL is the following:
- postgres try to lock `t1`'s row. Native postgres will raise and error message for `t1` because there is a conflict with another transaction. But YSQL doesn't perform the lock at this point and adds the lock intent into the buffer instead and returns result to postgres' upper layer that row was successfully locked
- postgres try to lock `t2`'s row (because it thinks that `t1`'s row is locked). YSQL adds the lock intent into the buffer
- At the end of statement processing YSQL flushes collected lock intends and sends all RPC to perform the read from multiple tables (`t1`, `t2`).
- At this point it is possible `t1` and `t2` has different t-server's leaders. And 2 independent RPC's will be sent simultaneously to different t-servers.
It is not deterministic what exact t-server will receive and process the request first. `t1` t-server leader or `t2` t-server leader. As a result YSQL will show error message with the `t1` or `t2` table with same probability. Native postgres always shows error message with `t1` table in described scenario.
Jira: DB-11445

Test Plan:
Jenkins

New unit tests are introduced

```
./yb_build.sh --gtest_filter PgMiscConflictsTest.ExplicitRowLockingSingleTableErrorMessage
./yb_build.sh --gtest_filter PgMiscConflictsTest.ExplicitRowLockingMultipleTablesErrorMessage
```

Reviewers: pjain, kramanathan

Reviewed By: kramanathan

Subscribers: smishra, ybase, yql

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D37821