[PLAT-15619]Task Details flicker if number of tasks exceeds 2000
Summary: /tasks api sends max 2000 tasks(unless otherwise configured) for a customer. Patch_For_tasks fetches tasks for the universe and add it to the list of tasks. but, again if this action is called it overrides the previous list of tasks.This causes ui to flicker. so, we are filtering the old tasks and adding the new tasks.

Test Plan: Tested manually

Reviewers: jmak, rmadhavan

Reviewed By: rmadhavan

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D38714