[PLAT-14766] Improve error message clarity for no available instances in a selected region
Summary:
When a provider does not have any available instances in a selected region, the error message prior to this diff is:
`No AZ found across regions: [<regionUuidList>]`.
This can be misleading because the customer may have availabilty zones in those regions. The underlying issue is that
those zones have no available instances.

This diff updates the error message to say `Couldn't find available nodes with type <instanceType> for given regions;
[<regionUuidList>]`.
This new error message will make it easier for a user to catch issues like an incorrectly selected instance type.

Test Plan:
- Create an on-prem provider with 2 regions and 2 instance types.
- Add instances to only one of the instance types.
- On the create universe page, select the above on-prem provider and select the instance type
  with instances attached.
- Change the region to the region without instances.
- Verify the error message is as expected.
{F294023}

Reviewers: yshchetinin

Reviewed By: yshchetinin

Subscribers: svc_phabricator, yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D38612