[BACKPORT 2024.2][PLAT-14597][PLAT-15338][PLAT-15523][PLAT-15636][PLAT-15526][PLAT-15603]
Summary:
[PLAT-14597] Fetch xCluster replication/DR configs with and without extra table info
**Context**
The YBA backend change in D37690 / b47e105e7f4d33dc250531dd5d22f38014225aea gives clients the ability to
skip the 'sync with DB' step when getting an xCluster replication/DR config.
This will allow users to recieve at least the DR/xCluster object we have stored in the YBA database even when
the universes are down and YBA is unable to sync with them.

Instead of getting an error on fetching the xCluster config, users will now be able to perform tasks like
switchover/failover with only UI elements like the replication tables list and edit tables/restart replication being
blocked.

**Changes**
- This diff makes the DR and xCluster list page skip the sync with DB step. This is because we don't show any table
details at all in the list page so we don't need to perform these rpc calls yet.
- On the DR / xCluster replication details page, we make a GET xCluster/DR call where we skip sync with DB and one
where we do sync with DB.
For UI elements that rely on table detail information, we show loading/error accordingly

[PLAT-15338] Improve formatting for very long universe names in DR participant cards

This diff changes the width for DR participant cards to min-width which allows the card to grow horizontally to
accomodate for very long univese names.

[PLAT-15523] Show tables tab for GET xCluster config without sync db option

If at least one of the xCluster config participant universes is down, then prior to this diff we would not show the xCluster table list as table details information would be missing.
This diff updates the YBA UI such that we are now able to see the tables with the understanding that the table name will be replaced with table UUID, db field will be missing and table status will be hidden.
Table level lag graph can be shown as well, but likewise the table details will be missing so we only show table UUID.

We are making this change simply to give the user more information to work with when they encounter this situation from the YBA UI.

[PLAT-15636] Add generated dr config name to pre-fill the create DR form

This is a quality of life improvement for setting up xCluster DR configs.
This diff pre-fills the create DR config form with a generated dr config name so that
users don't need to come up with one if they don't care about the name.

The DR config name remains editable so the user can choose to select their own if they wish.

[PLAT-15526][PLAT-15603] Update txn atomicity enabled classification to include DB scoped configs

Certain YBA UI elements are only applicable to xCluster configs with transactional atomicity enabled.
Prior to DB scoped xCluster configs being supported on YBA, we checked for xCluster config type of `Txn`
to determine if a particular config has transactional atomicity enabled.
Now that xCluster configs can also be db scoped, we need to classify both `Db` and `Txn` type xCluster configs
as transactional atomicity enabled configs.

This diff adds this classification logic as a helper function and replaces the `Txn` type check in applicable places
of the code.
We also hide the `Select unreplicated tables` button for DB scoped because we don't need to select individual tables for this xCluster config type.

Test Plan:
- Create xCluster replication config.
- Create xCluster DR config
- Edit tables / restart replication in xCluster replication config.
- Edit tables / restart replication in xCluster DR config
- Verify the UI still loads and only the replication tables section is loading/in error state.

Verify the DR participant card width grows as expected when given longer universe names.
Verify the overall UI layout is as expected.
{F287870}

Set up an xCluster config and stop all the nodes in one of the universes.
Verify that the xCluster table list is shown and is as described above.
{F293602}
{F293603}

Verify the user can create a DR config.
- Verify the user can use the generated DR config name or come up with their own
{F296985}

Verify that the consumer safe time lag and consumer safe time skew metrics are shown for db scoped xCluster configs.
Verify that the `active`/`standby` label is shown for db scoped xCluster configs in the xCluster replication tab.
{F296392}
{F296393}

Verify `Select unreplicated tables`  button is not rendered for db scoped xCluster DR

{F296993}

Reviewers: rmadhavan, hzare, vbansal, cwang, kkannan

Reviewed By: rmadhavan

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D39077