[#24173] DocDB, ASH: Fix rpc_request_id getting overwritten
Summary:
rpc_request_id is generated when a InboundCall object is
created, it shouldn't be updated from the protobuf.

This diff makes sure we don't update the rpc_request_id
of wait state from protobuf.

Note we need to pass the rpc_request_id to protobuf
because we use that to store samples in the circular
buffer
Jira: DB-13060

Test Plan: ./yb_build.sh --cxx-test wait_states-itest --gtest-filter WaitStateITest.UniqueRpcRequestId

Reviewers: amitanand

Reviewed By: amitanand

Subscribers: ybase

Differential Revision: https://phorge.dev.yugabyte.com/D38501