[BACKPORT 2024.1][PLAT-15456]: Allow users to drop failed DB from xCluster config
Summary:
In case, users fail to add a database to an xCluster config and the running and failed database(replication setup not done) has an equal or higher number of tables then our pre-checks fail assuming we are deleting all the tables in replication. But, in reality, we only remove tables with the replication setup not done.
Modified the check only to validate if the removed tables had replication setup done previously.

Original diff/commit: c0a93e2/D38750

Test Plan: Added local provider test

Reviewers: hzare, sanketh, cwang, #yba-api-review!

Reviewed By: cwang

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D38994