[PLAT-15636] Add generated dr config name to pre-fill the create DR form
Summary:
This is a quality of life improvement for setting up xCluster DR configs.
This diff pre-fills the create DR config form with a generated dr config name so that
users don't need to come up with one if they don't care about the name.

The DR config name remains editable so the user can choose to select their own if they wish.

Test Plan:
Verify the user can create a DR config.
- Verify the user can use the generated DR config name or come up with their own
{F296985}

Reviewers: rmadhavan, cwang, hzare, vbansal

Reviewed By: cwang

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D38769