[BACKPORT 2024.2][PLAT-15563] health check and alert errors when system clock is not in sync.
Summary:
Original commit: d42f4f520a6419571e0cded5cd0506463cf56cf3 / D38652
Updated the "node ntp service status" alert and health check to validate the
system clock being reported as "in sync". This is useful if it is not possible to reach
the remote ntp server, as no drift will be reported in these cases, but there is still
an issue.

Test Plan: validated health checks and alerts are thrown

Reviewers: yash.priyam, patnaik.balivada, muthu

Reviewed By: patnaik.balivada

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D38770