[#24246] YSQL: Fix out of bound array access in scan cost estimation for indexes on system columns
Summary:
This issue affects the new cost model which can be enabled using the
`yb_enable_base_scans_cost_model` GUC.

This bug may cause an out-of-bound memory access in an array and can lead to a
crash due to segmentation fault in some cases. This is caused by bug in
`yb_get_ybctid_width` function which computes the width of the `ybctid`.
This is used in computing the cost of an index scan operation. Following two
types of queries may be affected,
* Queries that access certain system catalogs with index on
  system columns
* Queries that access tables that have indexes on expressions

The `ybctid` contains the values of the index keys in an encoded form. To
compute the size of the `ybctid` we need to find the type and size of the
columns used in the index. For each index key, the
`(RelOptInfo*)index->indexkeys[]` contains the index of the column in the base
table. If the index key is an expression, then this contains value `0`. In case
of `pg_class` the primary index is on the `oid` column which is a system column
in PG11. `(RelOptInfo*)index->indexkeys[0]` contains value `-2` in this case.

To get the type of the column, we access 'attrs' array in `TupleDescData` which
describes the tuples in a table. However, it only contains data for the user
columns and not the system columns. In case of index on system columns or
index with expressions, we would access this array beyond the bounds. This is
fixed by checking the type only if the index key is a user column.
Jira: DB-13142

Test Plan: ./yb_build.sh --java-test 'org.yb.pgsql.TestPgEstimatedDocdbResultWidth'

Reviewers: tnayak, mihnea

Reviewed By: tnayak

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D38825