[BACKPORT 2024.2][#24116] CDCSDK: Store minimum start time of running txns in log segment footer
Summary:
**Backport description:**
No merge conflicts

**Original description:**
Original commit: 5f4f4cb7dfe163d1754ed5b5c0c89e976178e8bc / D36364
This diff adds a new proto field called `min_start_time_running_txns` to the wal segment footer. It will hold the value that represents the minimum start HT of all running txns at the time of closing of a log segment. If there are no running txns, this would represent the tablet leader safe time. It is being ensured that this value will always hold a valid HT value for closed segments.

This purpose of this field with respect to CDC is as follows:

  # Finding the stream safe time for the current WAL segment.
  ## This time represents the upper bound of all the records/txns in the segment that can be streamed and still guarantee that all records/txns with hybrid time < `min_start_time_running_txns` have been either streamed or are part of the current batch.
  ## Currently, we have to read all the WAL segments as we only have the stream safe time as of the GetChanges call. Now, `min_start_time_running_txns` will provide the stream safe time for each segment and hence, we can optimize WAL reads per GetChanges call.
  # Find out all the intents that can be cleaned up once the WAL segment is garbage collected.
  ## When a segment is garbage collected, it guarantees that all the transactions with APPLY records having OpId < `cdc_sdk_min_checkpoint_op_id` have been streamed by CDC. This implies that: The largest OpId in the segment < `cdc_sdk_min_checkpoint_op_id`.
  ## At this point, if we know the min start time of running txns (at that point in time when the segment was closed), we can safely clean up all the intents having hybrid time < `min_start_time_running_txns`.

Changes for the above mentioned purpose are being done in separate diffs.

**Changes:**
At the time of opening of a new segment, this field in the footer builder will be assigned `HybridTime::kInvalid`. The value of this field will not be changed as and when ops are being appended to the segment. At the time of closing of this segment which happens either during rollover or closing of log object, the value in the footer builder field will be updated via a callback (`min_start_ht_running_txns_callback_`) and written in the segment footer.

**min_start_ht_running_txns_callback_:**
  - This callback is initialised to `TabletPeer::GetMinStartHTRunningTxns()` during opening of log object as part of tablet bootstrap.
  - The callback is called right before closing of a log segment. It gets the minimum start time of all active transactions at that moment. If there does not exist any running transaction at that instant, then the tablet safe time will be assigned to the field. If unexpectedly, due to errors, we are not able to fetch the tablet safe time, `HybridTime::kInitial` will be assigned to this field.

Through this callback, it is guaranteed that the consistent stream safe time of a closed segment will always be a 'valid' hybrid time value.

**Upgrade / Rollback safety**
This diff introduces a new field to the `LogSegmentFooterPB`:
```
optional uint64 consistent_stream_safe_time = 6;
```
A new auto klocalPersisted flag is introduced - `store_min_start_ht_running_txns` which will be checked while modifying the `min_start_time_running_txns` field in the footer builder.

Note, the log segments generated before the upgrade will not contain this field in their footer. This should be handled by the reader of log segment footers. The getter function provided in the log_cache (`GetMinStartTimeRunningTxnsFromSegmentFooter()`) handles this by returning `HybridTime::kInitial` for such older segments.
Jira: DB-13006

Test Plan:
Modified existing log tests to assert presence of min_start_time_running_txns in footer of closed segments.

New tests:
./yb_build.sh --cxx-test tablet_peer-test --gtest_filter TabletPeerTest.TestMinStartTimeRunningTxnsOnLogSegmentRollover
./yb_build.sh --cxx-test ql-transaction-test --gtest_filter QLTransactionTest.LogSegmentRolloverWithRunningTxn

Reviewers: skumar, sumukh.phalgaonkar, rthallam, esheng, sergei

Reviewed By: skumar

Subscribers: ycdcxcluster, ybase, stiwary

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D38830