[#23769][2024.2] YSQL: Have a different pg regress result with connection manager
Summary:
When connection manager is enabled, we have to set the context of the client on the server before query execution and the context has to be reset post the query execution.

As part of setting the client context on the server connection, we run the SET statements with the session variables. This leads to additional query execution on the server.

Similarly during the reset phase, we run the query `RESET ROLE;SET SESSION AUTHORIZATION DEFAULT;RESET ALL` which can result in a different expected result set.

The changes here are a list of changes made to the expected results to accommodate the additional queries run by connection manager.
Jira: DB-12673

Test Plan:
Jenkins: .*TestPgRegressExtension.*|.*TestPgRegressThirdPartyExtensionsPgStatMonitor.*|.*TestPgRegressFeature.*, enable connection manager

Manually ran the above test classes locally without connection manager and verified that the tests are working as expected.

Reviewers: tfoucher, skumar, mihnea

Reviewed By: tfoucher

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D38888