[#20695] YSQL: Aggregate YB metric from parallel workers
Summary:
The parallel workers are instrumented and collect all the requested
metrics, including YB specific. However, thew are not aggregated with
the master. There is a function in the PG code to perform aggregation,
it should be extended to aggregate YB metrics as well.

The DocDB metrics collection when EXPLAIN ANALYZE's DEBUG is on wasn't
enabled in parallel workers. That part may need a follow up change, since
background worker can't tell if DEBUG is enabled or not, and always has to
collect these metrics. Probably we should introduce an analyze option flag for
that, something like INSTRUMENT_YBDEBUG.

Jira: DB-9699

Test Plan:
Check current tests.
In my experience DEBUG numbers vary from run to run, I'm not sure how to confirm that metrics are coming from the background workers, not just from the main worker.

Reviewers: esheng, kramanathan

Reviewed By: kramanathan

Subscribers: smishra, yql

Differential Revision: https://phorge.dev.yugabyte.com/D32836