[PLAT-15090] Add new API endpoint build Prometheus URL on YBA UI
Summary:
This diff adds a helper function for building the base Prometheus URL for a YBA instance using the values
defined in the runtime config.
In this diff, we fix an issue with the URL being generated for the metrics federation job which was using
the protocal of the YBA instance instead of the protocal of the Prometheus instance by making using of the
helper function we added.

Test Plan:
Verify the metrics URL correctly reflects what is set in the runtime config.
{F296340}
{F296974}
{F296973}
{F296975}
{F296976}

Reviewers: rmadhavan, amalyshev, muthu, #yba-api-review, sneelakantan

Reviewed By: rmadhavan, amalyshev, #yba-api-review

Subscribers: sneelakantan, yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D38182