[BACKPORT 2024.2][PLAT-15158][PLAT-15090] HA Improvements Backport
Summary:
[PLAT-15158 Update replication frequency tooltip

This diff adds a warning to the HA replication frequency field tooltip to inform users about triggering
the HA standby sync alert repeatedly if replication frequency is greater than the alert threshold.

[PLAT-15090] Add new API endpoint build Prometheus URL on YBA UI

This diff adds a helper function for building the base Prometheus URL for a YBA instance using the values
defined in the runtime config.
In this diff, we fix an issue with the URL being generated for the metrics federation job which was using
the protocal of the YBA instance instead of the protocal of the Prometheus instance by making using of the
helper function we added.

Test Plan:
- Verify the expected message appears on the HA replication frequency tooltip.
{F287109}

Verify the metrics URL correctly reflects what is set in the runtime config.
{F296340}
{F296974}
{F296973}
{F296975}
{F296976}

Reviewers: rmadhavan, muthu, amalyshev, sneelakantan

Reviewed By: amalyshev

Subscribers: sneelakantan, yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D39080