[#24147] YSQL: Introduce auth-backend for authentication in connection manager
Summary:
This revision introduces a special PG backend which will only be responsible for the **authentication** of logical connections via the connection manager. This mechanism will replace the existing auth-passthrough mechanism implemented in https://phorge.dev.yugabyte.com/D25292.

For a description of why this change is needed, please refer to the design document.

##### Role
The auth-backend will be responsible for two things:
1. Authenticate the user according to the hba rules defined by the administrator
2. Send the database-oid of the database the user is connecting to. This is because the connection-manager works with database OIDs instead of database names.
3. Send the default values of the GUC values marked with `GUC_REPORT` to the client

##### Lifecycle
In this new approach, the lifecycle of a logical connection at a high level will be as follows:
  - Logical connection created by the client with the connection manager
  - Connection manager will spawn an auth-backend and pass the required properties in the startup packet:
      - user
      - database
      - client's ip address. Used for matching correct hba rules
      - connection type (unencrypted or encrypted). Used for matching correct hba rules for SSL
  - The auth-backend will authenticate the connection and send the values of GUC parameters marked with `GUC_REPORT` to the client as per the protocol. The backend will then shut down
  - Upon successful auth, the connection manager chooses a backend from an existing pool of ready backends and assigns it to execute the user queries.

This revision **only includes the changes to implement the auth-backend**. The connection manager side changes to replace the usage of auth-passthrough with the auth-backend will be done in a follow-up revision.

##### Startup
To start an auth-backend, the client has to send the `yb_authonly` boolean parameter as `true` in the startup packet. This is very similar to how a [walsender backend](https://github.com/yugabyte/yugabyte-db/blob/d806031977c56c61848b041be068094706e6e2ac/src/postgres/src/backend/postmaster/postmaster.c#L2174-L2195) is started, which uses the `replication=database` or `replication=true` startup parameter.

##### Authentication
There are just a few changes in the authentication logic:
1. The connection type between the client and the conn-manager is hardcoded to `AF_INET` i.e. we assume that the client has a TCP connection with the conn-manager. Unix socket connection is currently unsupported by the conn-manager
2. The connection is treated as an SSL encrypted connection if the connection type passed in the startup packet is 'E'. This was already done for auth-passthrough. We also apply the same logic in auth-backend

##### Post Authentication
Post authentication:
1. database oid is fetched and sent to the client
2. The GUC parameters are sent to the client
3. The backend shuts down

##### Optimizations
Since the auth-backend has a limited role in comparison to a regular backend, we've disabled a few setup steps:
1. Active session history
2. Prefetching of pg_database and pg_authid tables. These tables are scanned and fetched during authentication and pre-fetching them isn't quite useful as the backend shuts down after that

There should be more optimization opportunities but they are left for the future.
Jira: DB-13035

Test Plan: Currently, the connection manager isn't using this auth-backend. So existing tests are sufficient to ensure that nothing is broken by this revision.

Reviewers: skumar, jason, myang, devansh.saxena

Reviewed By: myang

Subscribers: devansh.saxena, yql

Differential Revision: https://phorge.dev.yugabyte.com/D38459